合并Utility工具类
